requirements.txt
django>=3.1
djangorestframework>=3.12
python-magic>=0.4